create Service/Archive namespace and move ArchiveServices
